pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "traderbot"
version = "0.1.0"
description = "Educational trading bot with Alpaca integration"
readme = "README.md"
authors = [
    { name = "Trader Bot" }
]
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "pandas>=1.5",
    "numpy>=1.23"
]

[project.optional-dependencies]
live = ["alpaca-py>=0.13"]
dev = [
    "pytest>=7.4",
]

[project.scripts]
traderbot = "traderbot.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
